+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Version 0.1.1
 
-* Added functions: `fract`, `trunc`, `ceil`, `floor`, `sqrt`, `powi`.
+* Added functions: `recip`, `fract`, `trunc`, `ceil`, `floor`, `sqrt`, `powi`.
 * Added `%` and `%=` operators.
 
 ## Version 0.1.0